requirements.txt
requests~=2.32.3
streamlit~=1.41.1
Flask~=3.0.3
bs4~=0.0.2
beautifulsoup4~=4.13.1
openai~=1.60.1
python-dotenv~=1.0.1
pinecone~=5.4.2
langchain-openai